Changed to avoid compiler warning.
--HG--
extra : convert_revision : svn%3Afdd8eb12-d10e-0410-9acb-85c331704f74/trunk%403546

+++ dlib/statistics/statistics.h
@@ -224,7 +224,7 @@ namespace dlib
         long current_n (
         ) const
         {
-            return total_count;
+            return static_cast<long>(total_count);
         }
 
         template <typename EXP>
